Refactor newSendTxMsgServiceCmd
